Use ListMap instead of Map to avoid reordering
Case class parameters were being stored in a collection.immutable.Map,
which preserved their order for up to four elements, but after that the
order is based on hashes of the keys, and consequently the order appears
arbitrary.

This change replaces a few uses of `Map` with `ListMap`, notably in the
data macro and XML AST interface.

Quick review by @mielientiev.