Support for dictionaries
This will break a lot of code, as dynamic dereferencing is no longer the
default. To work around this, either import
rapture.data.dictionaries.dynamic._, or define a dictionary of keys,
like so:
```
implicit val dict = Dictionary.define("foo", "bar", "baz")
```
In this example, `foo`, `bar` and `baz` would be usable as lookup keys.